Fix some warnings from recent changes.

+++ codemp/qcommon/common.cpp
@@ -1762,7 +1762,7 @@ PrintCvarMatches
 static void PrintCvarMatches( const char *s ) {
 	char value[TRUNCATE_LENGTH] = {0};
 
-	if ( !Q_stricmpn( s, shortestMatch, strlen( shortestMatch ) ) ) {
+	if ( !Q_stricmpn( s, shortestMatch, (int)strlen( shortestMatch ) ) ) {
 		Com_TruncateLongString( value, Cvar_VariableString( s ) );
 		Com_Printf( S_COLOR_GREY"Cvar "S_COLOR_WHITE"%s = "S_COLOR_GREY"\""S_COLOR_WHITE"%s"S_COLOR_GREY"\""S_COLOR_WHITE"\n", s, value );
 	}
